Fixed Bug
Fixed a bug in which refreshing (and by extension, deleting) a chat removed the ability to reply to other chats via the Notification Reply Method (only allowing Accessibility Method which doesn't work for group chats and seems to be less stable). 

This was due to the methodology in which MainActivity.groupMeChats was completely rewritten and replaced in GroupQueryTask, without saving reply objects. 

Solved by copying over all available reply objects to new MainActivity.groupMeChats in GroupQueryTask.